tuning a bit

+++ codelearner_server/app/Http/Controllers/ProblemController.php
@@ -71,7 +71,8 @@ class ProblemController extends Controller implements HasMiddleware
     }
     
     public function update(Request $request, Problem $problem)
-    {
+    {   
+        Gate::authorize('modify', $problem);
         // validate the request
         $fields= $request->validate([
             'name' => 'string|max:255',
@@ -95,7 +96,8 @@ class ProblemController extends Controller implements HasMiddleware
     }
 
     public function destroy(Problem $problem)
-    {
+    {   
+        Gate::authorize('modify', $problem);
         // delete the problem
         $problem->delete();
